Attach eventd handlers in service container rather than in orc8r container
